Fix all outstanding shellcheck issues

+++ hack/bin/copy-charts.sh
@@ -14,14 +14,14 @@ CHART_DIST=$TOP_LEVEL/.local/charts
 # TODO sanity check folder must exist
 
 mkdir -p .local/charts
-rm -rf "$CHART_DIST/$CHART"
+rm -rf "${CHART_DIST:?}/$CHART"
 cp -r "$CHART_SOURCE/$CHART" "$CHART_DIST/"
 
 if [ -f "$CHART_SOURCE/$CHART/requirements.yaml" ]; then
   # very hacky bash, I'm sorry
   for subpath in $(grep "file://" "$CHART_SOURCE/$CHART/requirements.yaml" | awk '{ print $2 }' | xargs -n 1 | cut -c 8-)
   do
-    rm -rf "$CHART_DIST/$CHART/$subpath"
+    rm -rf "${CHART_DIST:?}/$CHART/$subpath"
     cp -r "$CHART_SOURCE/$CHART/$subpath" "$CHART_DIST/"
   done
 fi
